enable doing another test; disable forms after response
