+++ site/src/app/v2/test-eligibilite-juin/components/step-two-forms/AahCafForm.tsx
@@ -2,18 +2,17 @@ import Button from '@codegouvfr/react-dsfr/Button';
 import Input from '@codegouvfr/react-dsfr/Input';
 import { FormEvent, useRef, useState } from 'react';
 import {
+  AahCafInputsState,
   ConfirmResponseBody,
   ConfirmResponseError,
   SearchResponseBodyItem,
-  YoungCafInputsState,
 } from 'types/EligibilityTest';
 import { mapper } from './helper';
 import Alert from '@codegouvfr/react-dsfr/Alert';
+import FormButton from './FormButton';
 
-const initialInputsState: YoungCafInputsState = {
+const initialInputsState: AahCafInputsState = {
   recipientCafNumber: { state: 'default' },
-  recipientLastname: { state: 'default' },
-  recipientFirstname: { state: 'default' },
 };
 
 interface Props {
@@ -21,15 +20,16 @@ interface Props {
   onDataRecieved: (data: ConfirmResponseBody) => void;
 }
 
-const YoungCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
+const AahCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
   const formRef = useRef<HTMLFormElement>(null);
-  const [inputStates, setInputStates] = useState<YoungCafInputsState>(initialInputsState);
+  const [inputStates, setInputStates] = useState<AahCafInputsState>(initialInputsState);
   const [error, setError] = useState<string | null>();
+  const [isFormDisabled, setIsFormDisabled] = useState<boolean>(false);
 
-  const isFormValid = (formData: FormData): { isValid: boolean; states: YoungCafInputsState } => {
+  const isFormValid = (formData: FormData): { isValid: boolean; states: AahCafInputsState } => {
     let isValid = true;
 
-    const fieldNames = Object.keys(initialInputsState) as (keyof YoungCafInputsState)[];
+    const fieldNames = Object.keys(initialInputsState) as (keyof AahCafInputsState)[];
 
     const states = { ...initialInputsState };
 
@@ -70,8 +70,6 @@ const YoungCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
 
     const formData = new FormData(formRef.current!);
 
-    params.append('allocataireName', formData.get('recipientLastname') as string);
-    params.append('allocataireSurname', formData.get('recipientFirstname') as string);
     params.append('matricule', formData.get('recipientCafNumber') as string);
     params.append('id', eligibilityDataItem.id.toString());
     params.append('situation', eligibilityDataItem.situation);
@@ -106,6 +104,7 @@ const YoungCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
         notifyError(status, body as ConfirmResponseError);
       } else {
         onDataRecieved(body as ConfirmResponseBody);
+        setIsFormDisabled(true);
       }
     });
   };
@@ -123,30 +122,10 @@ const YoungCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
           }}
           state={inputStates.recipientCafNumber.state}
           stateRelatedMessage={inputStates.recipientCafNumber.errorMsg}
+          disabled={isFormDisabled}
         />
 
-        <Input
-          label="Nom de l’allocataire CAF*"
-          // hintText="Nom de la personne qui bénéficie des aides de la CAF ou la MSA"
-          nativeInputProps={{
-            name: 'recipientLastname',
-            placeholder: 'ex: Dupont',
-          }}
-          state={inputStates.recipientLastname.state}
-          stateRelatedMessage={inputStates.recipientLastname.errorMsg}
-        />
-
-        <Input
-          label="Prénom de l’allocataire CAF*"
-          // hintText="Nom de la personne qui bénéficie des aides de la CAF ou la MSA"
-          nativeInputProps={{ name: 'recipientFirstname', placeholder: 'ex: Marie' }}
-          state={inputStates.recipientFirstname.state}
-          stateRelatedMessage={inputStates.recipientFirstname.errorMsg}
-        />
-
-        <Button priority="primary" iconId="fr-icon-arrow-right-line" iconPosition="right">
-          Je vérifie mon éligibilité
-        </Button>
+        <FormButton isDisabled={isFormDisabled} />
       </form>
 
       {error && (
@@ -164,4 +143,4 @@ const YoungCafForm = ({ eligibilityDataItem, onDataRecieved }: Props) => {
   );
 };
 
-export default YoungCafForm;
+export default AahCafForm;
